Fix logic for tests
